pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alpha-sniper"
version = "0.1.0"
description = "Real-time stock news data pipeline for personal trading"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Data validation + config
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # MongoDB (raw zone)
    "beanie>=1.26.0,<2.0.0",
    "motor>=3.6.0",

    # PostgreSQL (clean zone)
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Task queue
    "taskiq[redis]>=0.11.0",
    "taskiq-redis>=1.0.0",

    # Redis
    "redis>=5.2.0",

    # WebSocket client
    "websockets>=14.0",

    # HTTP client
    "httpx>=0.28.0",

    # Retry logic
    "tenacity>=9.0.0",

    # Structured logging
    "structlog>=24.4.0",

    # XML/HTML parsing (SEC EDGAR)
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",

    # Feedparser (RSS)
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "faker>=28.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

ai = [
    "google-genai>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
